+++ MANIFEST.in
@@ -67,10 +67,10 @@ include psutil/arch/netbsd/disk.c
 include psutil/arch/netbsd/disk.h
 include psutil/arch/netbsd/mem.c
 include psutil/arch/netbsd/mem.h
+include psutil/arch/netbsd/proc.c
+include psutil/arch/netbsd/proc.h
 include psutil/arch/netbsd/socks.c
 include psutil/arch/netbsd/socks.h
-include psutil/arch/netbsd/specific.c
-include psutil/arch/netbsd/specific.h
 include psutil/arch/openbsd/cpu.c
 include psutil/arch/openbsd/cpu.h
 include psutil/arch/openbsd/disk.c
