+++ psutil/_psutil_bsd.c
@@ -96,7 +96,7 @@
     #include "arch/netbsd/cpu.h"
     #include "arch/netbsd/disk.h"
     #include "arch/netbsd/mem.h"
-    #include "arch/netbsd/specific.h"
+    #include "arch/netbsd/proc.h"
     #include "arch/netbsd/socks.h"
 
     #include <utmpx.h>
