+++ psutil/arch/netbsd/proc.c
@@ -13,7 +13,7 @@
 
 #include "../../_psutil_common.h"
 #include "../../_psutil_posix.h"
-#include "specific.h"
+#include "proc.h"
 
 
 #define PSUTIL_KPT2DOUBLE(t) (t ## _sec + t ## _usec / 1000000.0)
