+++ psutil/arch/netbsd/proc.h
