+++ setup.py
@@ -290,7 +290,7 @@ elif NETBSD:
             'psutil/arch/netbsd/cpu.c',
             'psutil/arch/netbsd/disk.c',
             'psutil/arch/netbsd/mem.c',
-            'psutil/arch/netbsd/specific.c',
+            'psutil/arch/netbsd/proc.c',
             'psutil/arch/netbsd/socks.c',
         ],
         define_macros=macros,
